Adds grid movement
Allows the use of the numpad to move in user-specified increments. Intended for use in marking polygon coordinates with 'space'.

Num0 will prompt the user for a new increment value (default 25 pixels).

Also added a ',' to the coordinate output in the console, to make it easy to paste the output into the map HTML file.